+++ CHANGELOG.md
@@ -12,11 +12,14 @@ The date format used is DD/MM/YYYY.
 
 - deleteAdventureById service
 - updateAdventureById service
+- DateTimePickerHF component
+- TimePickerHF component
 - EAS update
 
 ### Changed
 
 - Change the method (PUT -> PATCH) in updateAccountInformation service
+- Rename DateTimePickerHF component to DatePickerHF
 - Update the languageDectector plugin
 
 ## [[1.3.0]](https://github.com/willymateo/titi-app/pull/4) - 05/01/2023